college-api/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.1
djangorestframework==3.14.0
idna==3.6
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.4
numpy==1.26.3
psycopg2-binary==2.9.9
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.1.0
